requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
tensorflow==2.19.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
matplotlib==3.9.2
seaborn==0.13.2
Werkzeug==3.0.1
Jinja2==3.1.4
gunicorn==21.2.0
python-dotenv==1.0.0
joblib==1.4.2